@@ -25,81 +25,89 @@
   },
   "private": true,
   "dependencies": {
-    "@googlemaps/markerclusterer": "1.0.13",
+    "@googlemaps/markerclusterer": "^1.0.27",
     "@zxing/browser": "^0.1.1",
     "@zxing/library": "^0.19.1",
-    "async-mutex": "^0.1.3",
-    "bootstrap": "^4.3.1",
-    "core-js": "^3.6.5",
-    "document-register-element": "1.13.1",
-    "jquery": "^3.2.1",
+    "async-mutex": "^0.1.4",
+    "bootstrap": "^4.6.2",
+    "core-js": "^3.26.1",
+    "document-register-element": "^1.14.10",
+    "jquery": "^3.6.2",
     "lodash": "^4.17.21",
-    "query-string": "^6.14.0",
-    "rc-slider": "^8.6.13",
-    "react": "16.14.0",
+    "query-string": "^6.14.1",
+    "rc-slider": "^8.7.1",
+    "react": "^18.2.0",
     "react-barcode-reader": "^0.0.2",
     "react-cropper": "^2.1.8",
-    "react-datepicker": "^2.9.3",
-    "react-dom": "^16.8.2",
-    "react-google-autocomplete": "^1.1.2",
+    "react-datepicker": "^2.16.0",
+    "react-dom": "^18.2.0",
+    "react-google-autocomplete": "^1.3.0",
+    "react-is": "^18.2.0",
     "react-qr-reader": "^2.2.1",
-    "react-signature-canvas": "^1.0.1",
+    "react-signature-canvas": "^1.0.6",
     "reflect-metadata": "^0.1.13",
-    "regenerator-runtime": "0.13.7",
-    "swiper": "^6.7.5",
-    "tslib": "^2.1.0",
-    "typescript-collections": "^1.3.2"
+    "regenerator-runtime": "^0.13.11",
+    "swiper": "^6.8.4",
+    "tslib": "^2.4.1",
+    "typescript-collections": "^1.3.3"
   },
   "devDependencies": {
-    "@babel/core": "7.12.13",
-    "@babel/preset-env": "7.12.13",
-    "@babel/preset-react": "7.12.13",
-    "@babel/preset-typescript": "7.12.13",
-    "@nrwl/cli": "12.6.3",
-    "@nrwl/cypress": "12.6.3",
-    "@nrwl/eslint-plugin-nx": "12.6.3",
-    "@nrwl/jest": "12.6.3",
-    "@nrwl/linter": "12.6.3",
-    "@nrwl/react": "^12.6.3",
-    "@nrwl/tao": "12.6.3",
-    "@nrwl/web": "12.6.3",
-    "@nrwl/workspace": "12.6.3",
-    "@testing-library/react": "11.2.6",
-    "@types/bootstrap": "^4.3.1",
+    "@babel/core": "^7.20.5",
+    "@babel/preset-env": "^7.20.2",
+    "@babel/preset-react": "^7.18.6",
+    "@babel/preset-typescript": "^7.18.6",
+    "@nrwl/cli": "^15.2.1",
+    "@nrwl/cypress": "^15.2.1",
+    "@nrwl/eslint-plugin-nx": "^15.3.3",
+    "@nrwl/jest": "^15.2.1",
+    "@nrwl/linter": "^15.2.1",
+    "@nrwl/nx-cloud": "^15.0.2",
+    "@nrwl/react": "^15.3.3",
+    "@nrwl/web": "^15.3.3",
+    "@nrwl/workspace": "^15.2.1",
+    "@pmmmwh/react-refresh-webpack-plugin": "^0.5.10",
+    "@svgr/webpack": "^6.5.1",
+    "@testing-library/react": "^13.4.0",
+    "@types/bootstrap": "^4.6.2",
     "@types/googlemaps": "^3.43.3",
-    "@types/jest": "26.0.24",
-    "@types/jquery": "^3.2.1",
-    "@types/lodash": "^4.14.177",
-    "@types/node": "14.14.33",
-    "@types/rc-slider": "^8.6.4",
-    "@types/react": "16.8.3",
-    "@types/react-datepicker": "^2.3.0",
-    "@types/react-dom": "17.0.3",
+    "@types/jest": "^28.1.8",
+    "@types/jquery": "^3.5.14",
+    "@types/lodash": "^4.14.191",
+    "@types/node": "^18.11.15",
+    "@types/rc-slider": "^8.6.6",
+    "@types/react": "^18.0.26",
+    "@types/react-datepicker": "^2.11.1",
+    "@types/react-dom": "^18.0.9",
     "@types/react-qr-reader": "^2.1.4",
     "@types/react-signature-canvas": "^1.0.2",
     "@types/ts-nameof": "^4.2.1",
-    "@typescript-eslint/eslint-plugin": "~4.28.3",
-    "@typescript-eslint/parser": "~4.28.3",
-    "babel-jest": "27.0.6",
+    "@typescript-eslint/eslint-plugin": "^5.46.1",
+    "@typescript-eslint/parser": "^5.46.1",
+    "babel-jest": "^28.1.3",
     "babel-plugin-ts-nameof": "^4.2.1",
-    "cypress": "^7.3.0",
-    "dotenv": "~10.0.0",
-    "eslint": "7.22.0",
-    "eslint-config-prettier": "8.1.0",
-    "eslint-plugin-cypress": "^2.10.3",
-    "eslint-plugin-import": "2.22.1",
-    "eslint-plugin-jsx-a11y": "6.4.1",
-    "eslint-plugin-react": "7.23.1",
-    "eslint-plugin-react-hooks": "4.2.0",
+    "cypress": "^7.7.0",
+    "dotenv": "^10.0.0",
+    "eslint": "^8.29.0",
+    "eslint-config-prettier": "^8.5.0",
+    "eslint-plugin-cypress": "^2.12.1",
+    "eslint-plugin-import": "^2.26.0",
+    "eslint-plugin-jsx-a11y": "^6.6.1",
+    "eslint-plugin-react": "^7.31.11",
+    "eslint-plugin-react-hooks": "^4.6.0",
     "file-loader": "^6.2.0",
-    "jest": "27.0.3",
-    "prettier": "^2.3.1",
-    "ts-jest": "27.0.3",
+    "jest": "^28.1.3",
+    "jest-environment-jsdom": "^29.3.1",
+    "nx": "^15.2.1",
+    "prettier": "^2.8.1",
+    "react-refresh": "^0.14.0",
+    "react-test-renderer": "^18.2.0",
+    "ts-jest": "^28.0.8",
     "ts-nameof": "^5.0.0",
-    "ts-node": "~9.1.1",
-    "typescript": "3.8.2"
+    "ts-node": "^10.9.1",
+    "typescript": "^4.9.4",
+    "url-loader": "^4.1.1"
   },
   "peerDependencies": {
-    "jquery": "^3.2.1"
+    "jquery": "^3.6.2"
   }
 }
