@@ -1,9 +1,7 @@
-module.exports = {
+/* eslint-disable */
+
+export default {
   displayName: "ReApptor.TypeScript.Linq",
   preset: "../jest.preset.js",
-  transform: {
-    "^.+\\.[tj]sx?$": "babel-jest",
-  },
-  moduleFileExtensions: ["ts", "tsx", "js", "jsx"],
   coverageDirectory: "../coverage/libs/ReApptor.TypeScript.Linq"
 };
