@@ -3,13 +3,12 @@ import {ArrayExtensions} from "../index";
 ArrayExtensions();
 
 describe("skip", function() {
-    
     test("skip0from0", function () {
         const result: number[] = [].skip(0);
         expect(result).toEqual([]);
     });
     
-    test("skip-from0", function () {
+    test("skip-1from0", function () {
         const result: number[] = [].skip(-1);
         expect(result).toEqual([]);
     });
