@@ -30,6 +30,9 @@ export default class ArrayUtility {
     }
 
     public static chunk<T>(items: readonly T[], size: number): T[][] {
+        if (size < 1)
+            throw Error(`Size "${size}" out of range, must be at least 1 or greater.`);
+        
         const result: T[][] = [];
 
         const copy: T[] = [...items];
@@ -78,9 +81,12 @@ export default class ArrayUtility {
         for (let i: number = 0; i < length; i++) {
             const item: T = items[i];
             const valid: boolean = predicate(items[i], i);
-            if (valid) {
-                result.push(item);
+
+            if (!valid) {
+                break;
             }
+
+            result.push(item);
         }
         return result;
     }
