@@ -1,5 +1,5 @@
-import BasePageParameters from "./BasePageParameters";
 import ServerError from "./ServerError";
+import {BasePageParameters} from "../base/BasePage";
 
 export default interface IErrorPageParameters extends BasePageParameters {
     error: ServerError | null;
