@@ -2,8 +2,7 @@ import React from "react";
 import {Dictionary} from "typescript-collections";
 import {IService, ServiceType, Utility, ServiceProvider} from "@weare/reapptor-toolkit";
 import PageRoute from "../models/PageRoute";
-import {IBasePage, IBasePageConstructor} from "../base/BasePage";
-import BasePageParameters from "../models/BasePageParameters";
+import {BasePageParameters, IBasePage, IBasePageConstructor} from "../base/BasePage";
 
 export interface IPageDefinitions {
     createPageAsync(route: PageRoute): Promise<IBasePage>;
