@@ -1,61 +1,63 @@
-//Models:
+// Models:
 export * from "./models/FileModel";
-export * from "./models/GeoCoordinate";
-export * from "./models/GeoLocation";
-export * from "./models/TimeSpan";
-export * from "./models/IPagedList";
-export *  from "./models/ISelectListItem";
 export {default as FileModel} from "./models/FileModel";
+export * from "./models/GeoCoordinate";
 export {default as GeoCoordinate} from "./models/GeoCoordinate";
+export * from "./models/GeoLocation";
 export {default as GeoLocation} from "./models/GeoLocation";
+export * from "./models/TimeSpan";
 export {default as TimeSpan} from "./models/TimeSpan";
-export {default as IPagedList} from "./models/IPagedList";
-export {default as ISelectListItem}  from "./models/ISelectListItem";
+export * from "./models/IPagedList";
 
-//Utilities:
+// Interfaces:
+export * from "./models/IPagedList";
+export type {default as IPagedList} from "./models/IPagedList";
+export * from "./models/ISelectListItem";
+export type {default as ISelectListItem} from "./models/ISelectListItem";
+
+// Utilities:
 export * from "./ArrayUtility";
-export * from "./NumberUtility";
-export * from "./StringUtility";
-export * from "./BoolUtility";
-export * from "./DateUtility";
-export * from "./Utility";
-export * from "./HashCodeUtility";
 export {default as ArrayUtility} from "./ArrayUtility";
+export * from "./NumberUtility";
 export {default as NumberUtility} from "./NumberUtility";
+export * from "./StringUtility";
 export {default as StringUtility} from "./StringUtility";
+export * from "./BoolUtility";
 export {default as BoolUtility} from "./BoolUtility";
+export * from "./DateUtility";
 export {default as DateUtility} from "./DateUtility";
+export * from "./Utility";
 export {default as Utility} from "./Utility";
+export * from "./HashCodeUtility";
 export {default as HashCodeUtility} from "./HashCodeUtility";
 
-//Extensions:
+// Extensions:
 export * from "./extensions/DateExtensions";
 export * from "./extensions/StringExtensions";
 export * from "./extensions/ArrayExtensions";
 export * from "./extensions/NumberExtensions";
 
-//Providers:
+// Providers:
 export * from "./providers/BaseTransformProvider";
-export * from "./providers/BaseEnumProvider";
-export * from "./providers/TypeResolver";
-export * from "./providers/ServiceProvider";
-export * from "./providers/TypeConverter";
-export * from "./providers/StringConverter";
 export {default as BaseTransformProvider} from "./providers/BaseTransformProvider";
+export * from "./providers/BaseEnumProvider";
 export {default as BaseEnumProvider} from "./providers/BaseEnumProvider";
+export * from "./providers/TypeResolver";
 export {default as TypeResolver} from "./providers/TypeResolver";
+export * from "./providers/ServiceProvider";
 export {default as ServiceProvider} from "./providers/ServiceProvider";
+export * from "./providers/TypeConverter";
 export {default as TypeConverter} from "./providers/TypeConverter";
+export * from "./providers/StringConverter";
 export {default as StringConverter} from "./providers/StringConverter";
 
-//Helpers:
+// Helpers:
 export * from "./helpers/PwaHelper";
 export {default as PwaHelper} from "./helpers/PwaHelper";
 
-//Localizer:
+// Localizer:
 export * from "./localization/BaseLocalizer";
 export {default as BaseLocalizer} from "./localization/BaseLocalizer";
-export {Language} from "./localization/BaseLocalizer";
 
 //Other:
 export * from "./AthenaeumConstants";
