@@ -1,7 +1,6 @@
-import {ApplicationContext, BasePage} from "@weare/reapptor-react-common";
+import {ApplicationContext, BasePage, TBasePageParameters} from "@weare/reapptor-react-common";
 
-
-export default abstract class AnonymousPage<TParams = {}, TState = {}>
+export default abstract class AnonymousPage<TParams extends TBasePageParameters = {}, TState = {}>
     extends BasePage<TParams, TState, ApplicationContext> {
 
     protected get typedParameters(): TParams | null {
