@@ -4,16 +4,15 @@ import {BasePageParameters, ch, PageRouteProvider} from "@weare/reapptor-react-c
 import AnonymousPage from "../AnonymousPage";
 import Localizer from "../../../localization/Localizer";
 
-
 export interface IAnonymousParameters extends BasePageParameters {
     hello: string;
     world: object;
 }
 
 export default class AnonymousTestWithParameters extends AnonymousPage<IAnonymousParameters> {
-
+    
     public getTitle(): string {
-        return nameof(AnonymousTestWithParameters);
+        return nameof<AnonymousTestWithParameters>();
     }
 
     private async updateIdAsync(): Promise<void> {
@@ -41,7 +40,7 @@ export default class AnonymousTestWithParameters extends AnonymousPage<IAnonymou
     public render(): React.ReactNode {
 
         const localizerKey: string = `PageRoutes.${nameof(AnonymousTestWithParameters)}`;
-
+        
         console.log(
             [this.getTitle(), nameof(this.render)].join("."),
             this.route,
